requirements.txt
customtkinter
ctk
pyperclip
asyncio
CTkMessagebox
future
Pillow
bitarray
numpy